requirements.txt
exifread
python-telegram-bot